+++ asterisk/rootfs/etc/cont-init.d/register-ingress-entry.sh
@@ -5,12 +5,17 @@
 
 # shellcheck shell=bash
 
-function api_call {
-  curl -sS -H "Authorization: Bearer $SUPERVISOR_TOKEN" -H "content-type: application/json" "$@"
-}
+if [[ -z "${SUPERVISOR_TOKEN:-}" ]]; then
+    exit 0
+fi
 
-if [[ -n "${SUPERVISOR_TOKEN:-}" ]]; then
-    INGRESS_ENTRY=$(api_call -X GET http://supervisor/addons/self/info | jq -r .data.ingress_entry)
-    bashio::log.info ingress_entry: $INGRESS_ENTRY
-    api_call -X POST http://supervisor/core/api/states/text.asterisk_addon_ingress_entry -d '{"state": "'${INGRESS_ENTRY}'"}'
+if bashio::config.true 'register_ingress_entry'; then
+    ingress_entry=$(
+        curl -fsSL -H "Authorization: Bearer ${SUPERVISOR_TOKEN}" http://supervisor/addons/self/info |
+            jq -er .data.ingress_entry
+    )
+    bashio::log.debug "ingress_entry: ${ingress_entry}"
+    curl -fsSL -H "Authorization: Bearer ${SUPERVISOR_TOKEN}" \
+        -X POST http://supervisor/core/api/states/text.asterisk_addon_ingress_entry \
+        -d '{"state": "'"${ingress_entry}"'"}'
 fi